% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogenelist.R
\name{gogenelist}
\alias{gogenelist}
\title{Generates gene list for selected chromatin interactions}
\usage{
gogenelist(datafile, proximalmark, distalmark, gene.symbol = FALSE,
  mart = "ensembl", martset = "hsapiens_gene_ensembl", gbuild = 37)
}
\arguments{
\item{datafile}{The GRanges object output from the overlap function}

\item{proximalmark}{The name of the mark in the GRanges object which you are
interested in obtaining nearest gene expression data from}

\item{distalmark}{The name of the mark in the GRanges object interacting
with the proximalmark}

\item{gene.symbol}{If TRUE, will ouput gene list using official gene symbol.
If FALSE, gene list output will use Ensembl ID}

\item{mart}{BiomaRt used for pulling official gene name}

\item{martset}{The specific mart or dataset used for gene information}

\item{gbuild}{The genome version of mart to be used}
}
\value{
Returns a gene list sorted by gene expression levels of all genes
near the proximal mark that are interacting with the distal mark
}
\description{
Generates gene list for selected chromatin interactions
}
\examples{
gogenelist(datafile = overlapoutput, proximalmark = "PAR", distalmark = "EAR")
}
