% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicageshiny.R
\name{hicageshiny}
\alias{hicageshiny}
\title{Launch HiCAGE Shiny application}
\usage{
hicageshiny()
}
\value{
When run, function will launch the HiCAGE Shiny application - a GUI
for user-friendly implementation of HiCAGE
}
\description{
Launch HiCAGE Shiny application
}
