% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Combine chromosome conformation capture data with segmentation and RNA data}
\usage{
overlap(hicfile, segmentfile, rnafile, mart = "ensembl",
  martset = "hsapiens_gene_ensembl", gbuild = 37, hic.columns = c(1:6, 8),
  segment.columns = c(1:5), rna.columns = c(1, 7))
}
\arguments{
\item{hicfile}{The chromosome conformation capture datafile}

\item{segmentfile}{The segmentation datafile from StateHub}

\item{rnafile}{The RNA gene expression datafile containing Ensembl IDs and
FPKM or TPM normalized expression values}

\item{mart}{BiomaRt used for pulling gene name and location data}

\item{martset}{The specific mart or dataset used for genome information}

\item{gbuild}{The genome version of mart to be used}

\item{hic.columns}{Columns from the chromsome conformation capture data
that contains, in order, 'left chromosome', 'left start', 'left end', 'right
chromosome', 'right start', 'right end', 'interaction score'}

\item{manualselect}{If TRUE, allows for manual selection of chromosome
conformation capture columns to be used for analysis}
}
\value{
Returns a GRanges with each chromsome conformation capture segment
annotated with a prioritized segmentation mark, the nearest gene to the
segmentation mark and the gene's FPKM expression data
}
\description{
Combine chromosome conformation capture data with segmentation and RNA data
}
\examples{
overlap("hic_chr20.txt", "segment_chr20.bed", "rna_chr20.tsv")
example <- overlap(hicfile = "data/hic_chr20.txt", segmentfile = "data/segment_chr20.bed", rnafile = "data/rna_chr20.tsv")
example <- overlap(hicfile = "data/hic.txt", segmentfile = "data/segment.bed", rnafile = "data/rna.tsv")
}
