% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotup.R
\name{plotup}
\alias{plotup}
\title{Generates UpSetR-style plot from overlap data}
\usage{
plotup(datafile, ...)
}
\arguments{
\item{datafile}{The object output from the overlap function}

\item{...}{Additional arguments passed to upset function from UpSetR package}
}
\value{
Returns an UpSetR plot from overlap data
}
\description{
Generates UpSetR-style plot from overlap data
}
\examples{
hic_chr20 <- system.file("extdata", "hic_chr20.txt", package = "HiCAGE")
segment_chr20 <- system.file("extdata", "segment_chr20.bed",
    package = "HiCAGE")
rna_chr20 <- system.file("extdata", "rna_chr20.tsv", package = "HiCAGE")
overlapoutput <- overlap(hicfile = hic_chr20,
                         segmentfile = segment_chr20,
                         rnafile = rna_chr20)
plotup(datafile = overlapoutput)
}
