% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleplot.R
\name{circleplot}
\alias{circleplot}
\title{Generate circos plots from overlap data output}
\usage{
circleplot(datatable, plot.subset = FALSE, display.legend = TRUE,
  hic.legend = "Avg. Hi-C Score", hic.range = c(0, 130),
  rna.legend = "Avg. log(FPKM)", rna.range = c(0, 3.2),
  circos.color = NULL, ...)
}
\arguments{
\item{datatable}{The GRanges datatable generated by overlap function}

\item{plot.subset}{Allows for plotting interactions with Hi-C scores only}

\item{display.legend}{Adds legend to circos plot}

\item{hic.legend}{Title for Hi-C legend}

\item{hic.range}{A vector representing the low and high values for the legend
of Hi-C values}

\item{rna.legend}{Title for RNA legend}

\item{rna.range}{A vector representing the low and high values for the legend
of RNA values}

\item{circos.color}{Passes colors to chord diagram for each group}

\item{...}{Addtional arguments passed to chordDiagram function}
}
\value{
Generates a chord diagram containing circluar tracks with an inner
track displaying  average Hi-C interaction score and an inner track with
displaying average RNA-seq log(FPKM) values for each segment in the plot
}
\description{
Generate circos plots from overlap data output
}
\examples{
hic_chr20 <- system.file("extdata", "hic_chr20.txt", package = "HiCAGE")
segment_chr20 <- system.file("extdata", "segment_chr20.bed",
    package = "HiCAGE")
rna_chr20 <- system.file("extdata", "rna_chr20.tsv", package = "HiCAGE")
example <- overlap(hicfile = hic_chr20,
         segmentfile = segment_chr20,
         rnafile = rna_chr20,
         rna.columns = c(1, 7))
circleplot(example)
}
