% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleplot.R
\name{circleplot}
\alias{circleplot}
\title{Generate circos plots from overlap data output}
\usage{
circleplot(datatable, plot.subset = FALSE, display.legend = TRUE,
  hic.legend = "Avg. Hi-C Score", hic.range = c(0, 130),
  rna.legend = "Avg. log(FPKM)", rna.range = c(0, 3.2),
  circos.color = NULL, ...)
}
\arguments{
\item{datatable}{The GRanges datatable generated by overlap function}

\item{plot.subset}{Allows for plotting interactions with Hi-C scores only}

\item{display.legend}{Adds legend to circos plot}

\item{hic.legend}{Title for Hi-C legend}

\item{rna.legend}{Title for RNA legend}

\item{circos.color}{Passes colors to chord diagram for each group}
}
\description{
Generate circos plots from overlap data output
}
