% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Combine chromosome conformation capture data with segmentation and RNA data}
\usage{
overlap(hicfile, segmentfile, rnafile, mart = "ensembl",
  martset = "hsapiens_gene_ensembl", gbuild = 37, hic.columns = c(1:6, 8),
  segment.columns = c(1:5), rna.columns = c(1, 7))
}
\arguments{
\item{hicfile}{The chromosome conformation capture datafile}

\item{segmentfile}{The segmentation datafile from StateHub}

\item{rnafile}{The RNA gene expression datafile containing Ensembl IDs and
FPKM or TPM normalized expression values}

\item{mart}{BiomaRt used for pulling gene name and location data}

\item{martset}{The specific mart or dataset used for genome information}

\item{gbuild}{The genome version of mart to be used}

\item{hic.columns}{Columns from the chromsome conformation capture data file
that contain, in order, 'left chromosome', 'left start', 'left end', 'right
chromosome', 'right start', 'right end', 'interaction score'}

\item{segment.columns}{Columns from the segmentation data file that contain,
in order, 'Chromosome', 'Start', 'End', 'Mark', 'Score'}

\item{rna.columns}{Columns from the RNA-seq data file that contain, in order,
'GeneID', 'FPKM'}
}
\value{
Returns a GRanges with each chromsome conformation capture segment
annotated with a prioritized segmentation mark, the nearest gene to the
segmentation mark and the gene's FPKM expression data
}
\description{
Combine chromosome conformation capture data with segmentation and RNA data
}
\examples{
hic_chr20 <- system.file("extdata", "hic_chr20.txt", package = "HiCAGE")
segment_chr20 <- system.file("extdata", "segment_chr20.bed", package = "HiCAGE")
rna_chr20 <- system.file("extdata", "rna_chr20.tsv", package = "HiCAGE")
example <- overlap(hicfile = hic_chr20,
                   segmentfile = segment_chr20,
                   rnafile = rna_chr20)
}
